Refactor - create grid.rs file for grid type
